mini-projects/summariser/requirements.txt
langchain
langchain_community
OpenAI
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras